Use short for builtin. Alternatively we could use _Float16, but then need to enable that for the platform.

+++ clang/test/CodeGen/builtins-wasm.c
@@ -802,7 +802,7 @@ f32x4 relaxed_dot_bf16x8_add_f32_f32x4(u16x8 a, u16x8 b, f32x4 c) {
   // WEBASSEMBLY-NEXT: ret
 }
 
-float load_f16_f32(int *addr) {
+float load_f16_f32(short *addr) {
   return __builtin_wasm_loadf16_f32(addr);
   // WEBASSEMBLY: call float @llvm.wasm.loadf16.f32(ptr %{{.*}})
 }
